refactor of core module.
