+++ src/numlenology/__init__.py
@@ -1,50 +1,17 @@
 from collections import defaultdict
 
-import graphviz as gv
 import numpy as np
 from num2words import CONVERTER_CLASSES, num2words
 from scipy.sparse import csgraph
 
-from numlenology.utils import float_range, rank_keys
-from numlenology.utils.color import fade_color
 
+SUPPORTED_LANGS = sorted(set(lang[:2] for lang in CONVERTER_CLASSES.keys()))
 
-LANGS = sorted(set(lang[:2] for lang in CONVERTER_CLASSES.keys()))
-
-NODE_COLOR_LEAF = "#009000bb"
-NODE_COLOR_ROOT = "#000000bb"
-NODE_SIZE_MIN = 0.4
-NODE_SIZE_MAX = 0.9
-
-"""
-# connected components.
-cc_num, cc_labels = csgraph.connected_components(m)
-if cc_num > 1:
-    for label in range(cc_num):
-        cc_nodes = np.where(cc_labels == label)[0]
-        # cc_nodes.mean()
-        # cc_nodes.std()
-        # cc_nodes.sum()
-        # len(cc_nodes)
-
-
-node = 10
-
-# with axis=0 we calculate for each node X how many times we jumped from any node to X
-# when we built the matrix. then we get the value for this `node`.
-degree = m.sum(axis=0)[node]
-
-# get chain until loop.
-dfo = csgraph.depth_first_order(m, node, return_predecessors=False)
-# node that "closes" the loop, the only child node of `dfo[-1]`.
-ncl = m[dfo[-1]].argmax()
-# depth is equal to the number of nodes in `dfo` until we reach the end loop.
-depth = np.where(dfo == ncl)[0][0]
-"""
 
 def recursive_build_chain(lang, matrix, i, visited):
     visited.append(i)
-    j = len(num2words(i, lang=lang))
+    w = num2words(i, lang=lang)
+    j = len(w)
     if j in visited:
         # add the last edge to "close" the loop.
         matrix[i][j] += 1
@@ -54,17 +21,15 @@ def recursive_build_chain(lang, matrix, i, visited):
         return recursive_build_chain(lang, matrix, j, visited)
 
 
-def build_graph_matrix_rec(bound=10, lang="es"):
-    bound += 1  # from 0 to `bund` included.
-    matrix = np.zeros([bound, bound], dtype=np.int16)
+def build_graph_matrix_rec(bound, lang):
+    matrix = np.zeros([bound, bound], dtype=np.uint16)
     for i in range(bound):
         recursive_build_chain(lang, matrix, i, [])
     return matrix
 
 
-def build_graph_matrix(bound=10, lang="es"):
-    bound += 1  # from 0 to `bund` included.
-    matrix = np.zeros([bound, bound], dtype=np.int16)
+def build_graph_matrix(bound, lang):
+    matrix = np.zeros([bound, bound], dtype=np.uint16)
     for i in range(bound):
         visited = [i]
         while (j := len(num2words(i, lang=lang))) not in visited:
@@ -78,7 +43,7 @@ def build_graph_matrix(bound=10, lang="es"):
 
 # old methods.
 
-def build_single_chain(n, lang="en"):
+def build_single_chain(n, lang):
     """
     Build a chain with the following relationship:
         `n` in the language `lang` is `x` letters long.
@@ -115,7 +80,7 @@ def build_single_chain(n, lang="en"):
     return visited, end_loop
 
 
-def build_graph(bound=100, lang="en"):
+def build_graph(bound, lang):
     """
     Call `build_single_chain` for each number from 0 to `bound` included
     and return the following tuple:
@@ -130,7 +95,7 @@ def build_graph(bound=100, lang="en"):
     degree_by_node = defaultdict(int)
     depth_by_node = defaultdict(int)
 
-    for n in range(bound + 1):
+    for n in range(bound):
         visited, end_loop = build_single_chain(n, lang)
         depth_by_node[n] = len(visited) - len(end_loop)
 
@@ -153,48 +118,38 @@ def build_graph(bound=100, lang="en"):
     return connected_comp, graph_edges, degree_by_node, depth_by_node
 
 
-def build_graphviz(bound, lang):
-    """Call `build_graph` and return a graphviz.Digraph object."""
-    _, graph_edges, degree_by_node, depth_by_node = build_graph(bound, lang)
+class Graph:
+    """Class that contains the Numlenology representation for a given language."""
 
-    depths = rank_keys(depth_by_node, reverse=True)
-    color_scale = fade_color(
-        NODE_COLOR_LEAF,
-        NODE_COLOR_ROOT,
-        1 + max(depths.values()),
-    )
+    def __init__(self, lang, bound=100):
+        bound += 1  # work from 0 to `bund` included.
 
-    degrees = rank_keys(degree_by_node, reverse=True)
-    size_scale = float_range(
-        NODE_SIZE_MAX,
-        NODE_SIZE_MIN,
-        1 + max(degrees.values()),
-    )
+        m = build_graph_matrix_rec(bound, lang)
+        self.matrix = m
 
-    # build graph. graphviz requires objects to be cast to str.
-    g = gv.Digraph(
-        engine="neato",
-        node_attr={
-            "style": "filled",
-            "color": NODE_COLOR_LEAF,
-            "fixedsize": "true",
-            "width": str(NODE_SIZE_MIN),
-            "height": str(NODE_SIZE_MIN),
-        },
-        graph_attr={
-            "outputorder": "nodesfirst",
-        },
-    )
+        # connected components.
+        self.ccs = {}
+        cc_num, cc_labels = csgraph.connected_components(m)
+        for label in range(cc_num):
+            cc_nodes = np.where(cc_labels == label)[0]
+            self.ccs[label] = cc_nodes
+        # if cc_num > 1:
+            # cc_nodes.mean()
+            # cc_nodes.std()
+            # cc_nodes.sum()
+            # len(cc_nodes)
 
-    for node, degree in degrees.items():
-        g.node(
-            str(node),
-            color=str(color_scale[depths[node]]),
-            width=str(size_scale[degree]),
-            height=str(size_scale[degree]),
-        )
+        # with axis=0 we calculate for each node X how many times we jumped from any node to X
+        # when we built the matrix. then we get the value for this `node`.
+        self.gbn = self.degree_by_node = m.sum(axis=0)
 
-    for e in graph_edges:
-        g.edge(*map(str, e))
-
-    return g
+        # get chain until loop.
+        depth_by_node = np.zeros([bound], dtype=np.uint16)
+        for node in range(bound):
+            dfo = csgraph.depth_first_order(m, node, return_predecessors=False)
+            # node that "closes" the loop, the only child node of `dfo[-1]`.
+            ncl = m[dfo[-1]].argmax()
+            # depth is equal to the number of nodes in `dfo` until we reach the end loop.
+            depth = np.where(dfo == ncl)[0][0]
+            depth_by_node[node] = depth
+        self.dbn = self.depth_by_node = depth_by_node
